+++ src/buildInfo.ts
@@ -1,0 +1,1 @@
+export const __BUILD_DATE__ = '2025-07-08T18:25:12Z';